Seed phrase retrivier moved to dedicate reusable component
