+++ src/pages/Setting/ExportSeedPhrase/index.tsx
@@ -1,24 +1,15 @@
+import { useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import images from 'src/images';
 import styled from 'styled-components';
-import Button from 'src/components/Buttons';
-import { useEffect, useState } from 'react';
-import { useSelector } from 'react-redux';
-import { hash as kadenaHash } from '@kadena/cryptography-utils';
 import { toast } from 'react-toastify';
 import Toast from 'src/components/Toast/Toast';
-import { getLocalSeedPhrase } from 'src/utils/storage';
-import { BaseTextInput, InputError } from 'src/baseComponent';
-import { decryptKey } from 'src/utils/security';
 import { ReactComponent as AlertIcon } from 'src/images/icon-alert.svg';
 import { NavigationHeader } from 'src/components/NavigationHeader';
 import { DivFlex, SecondaryLabel } from 'src/components';
 import { Warning } from 'src/pages/SendTransactions/styles';
 import { ActionList } from 'src/components/ActionList';
-
-const Body = styled.div`
-  height: 170px;
-`;
+import { SeedPhraseRetrivier } from 'src/components/SeedPhraseRetrivier';
 
 const Wrapper = styled.div`
   padding: 0 20px;
@@ -38,15 +29,6 @@ export const SPWrapper = styled.div`
   text-align: center;
 `;
 
-const LockImage = styled.img`
-  width: 28px;
-  height: 35px;
-`;
-
-const CustomButton = styled.div`
-  margin-top: 20px;
-`;
-
 const SPText = styled.div`
   ${(props) => (props.isBlur ? 'filter: blur(4px);' : '')};
   color: ${({ theme }) => theme.text.primary};
@@ -56,33 +38,12 @@ const SPText = styled.div`
   font-size: 24px;
 `;
 
-const DivError = styled.div`
-  margin-top: 10px;
-`;
-const defaultText = '***** ***** ***** ***** ***** ***** ***** ***** ***** ***** ***** *****';
-
 const ExportSeedPhrase = () => {
   const history = useHistory();
-  const rootState = useSelector((state) => state);
-  const { passwordHash } = rootState.extensions;
-  const [seedPhraseHash, setSeedPhraseHash] = useState('');
-  const [passwordInput, setPasswordInput] = useState('');
-  const [isHiddenSP, setIsHiddenSP] = useState(true);
-  const [sP, setSP] = useState('');
-  const [isErrorEmpty, setErrorEmpty] = useState(false);
-  const [isErrorVerify, setErrorVerify] = useState(false);
-
-  useEffect(() => {
-    getLocalSeedPhrase(
-      (hash) => {
-        setSeedPhraseHash(hash);
-      },
-      () => {},
-    );
-  }, []);
+  const [seedPhrase, setSeedPhrase] = useState('');
 
   const onCopy = () => {
-    navigator.clipboard.writeText(sP);
+    navigator.clipboard.writeText(seedPhrase);
     toast.success(<Toast type="success" content="Copied!" />);
   };
 
@@ -103,19 +64,11 @@ const ExportSeedPhrase = () => {
     return result;
   };
 
-  const showSP = (hashInput) => {
-    setIsHiddenSP(false);
-    try {
-      const plainSeedPhrase = decryptKey(seedPhraseHash, hashInput);
-      setSP(plainSeedPhrase);
-    } catch (err) {
-      console.error('Invalid hash');
-    }
-  };
+  const onSeedPhraseRetrivied = (seedPhrase) => setSeedPhrase(seedPhrase);
 
   const onDownload = () => {
     const element = document.createElement('a');
-    element.setAttribute('href', `data:text/plain;charset=utf-8,${encodeURIComponent(sP)}`);
+    element.setAttribute('href', `data:text/plain;charset=utf-8,${encodeURIComponent(seedPhrase)}`);
     element.setAttribute('download', 'secret-recovery-phrase');
     element.style.display = 'none';
     document.body.appendChild(element);
@@ -123,34 +76,12 @@ const ExportSeedPhrase = () => {
     document.body.removeChild(element);
   };
 
-  const onChangeInput = (e) => {
-    setPasswordInput(e.target.value);
-    if (e.target.value) {
-      setErrorEmpty(false);
-      setErrorVerify(false);
-    } else {
-      setErrorVerify(false);
-    }
-  };
-  const handleVerifyPassword = () => {
-    if (!passwordInput) {
-      setErrorEmpty(true);
-    } else {
-      const hashInput = kadenaHash(passwordInput);
-      const isValid = hashInput === passwordHash;
-      if (isValid) {
-        showSP(hashInput);
-      } else {
-        setErrorVerify(true);
-      }
-    }
-  };
   return (
     <Wrapper>
       <NavigationHeader title="Export Recovery Phrase" onBack={goBack} />
-      {!isHiddenSP && (
+      {seedPhrase ? (
         <SPWrapper>
-          <SPText>{sP ? encryptText(sP) : defaultText}</SPText>
+          <SPText>{encryptText(seedPhrase)}</SPText>
           <DivFlex flexDirection="column" alignItems="center" paddingTop="30px" style={{ borderTop: '1px solid #DFDFED' }}>
             <SecondaryLabel fontWeight={500} fontSize={14}>
               Your Secret Recovery Phrase makes it easy to back up and restore your account.
@@ -171,38 +102,9 @@ const ExportSeedPhrase = () => {
             ]}
           />
         </SPWrapper>
+      ) : (
+        <SeedPhraseRetrivier onSuccess={onSeedPhraseRetrivied} />
       )}
-      {isHiddenSP && (
-        <DivFlex flexDirection="column" justifyContent="center" alignItems="center" padding="50px 0px">
-          <LockImage src={images.settings.iconLockOpen} alt="lock" />
-          <SecondaryLabel fontWeight={500}>Enter your password to continue</SecondaryLabel>
-        </DivFlex>
-      )}
-      <Body>
-        {isHiddenSP && (
-          <>
-            <BaseTextInput
-              inputProps={{ placeholder: 'Input password', type: 'password' }}
-              title="Password"
-              height="auto"
-              onChange={onChangeInput}
-              onKeyPress={(event) => {
-                if (event.key === 'Enter') {
-                  handleVerifyPassword();
-                }
-              }}
-              typeInput="password"
-            />
-            <DivError>
-              {isErrorEmpty && <InputError marginTop="0">This field is required.</InputError>}
-              {isErrorVerify && <InputError marginTop="0">Invalid Passwords.</InputError>}
-            </DivError>
-            <CustomButton>
-              <Button size="full" variant="primary" onClick={handleVerifyPassword} isDisabled={!passwordInput} label="Continue" />
-            </CustomButton>
-          </>
-        )}
-      </Body>
     </Wrapper>
   );
 };
